move files around and get customize query aligned
