+++ query-connector/src/app/query/components/resultsView/resultsView.module.css
@@ -18,6 +18,11 @@
     "Arial", "sans-serif" !important;
   font-size: 1.46rem !important;
   line-height: 1.1 !important;
-  padding: 0.75rem 0;
+  padding-bottom: 0.75rem;
   margin-bottom: 0.5rem !important;
+}
+
+.resultsBannerContent {
+  display: flex;
+  align-items: center;
 }
