+++ query-connector/src/app/query/components/resultsView/tableComponents/DiagnosticReportTable.tsx
@@ -20,7 +20,7 @@ const DiagnosticReportTable: React.FC<DiagnosticReportTableProps> = ({
   diagnosticReports,
 }) => {
   return (
-    <Table>
+    <Table className="margin-top-0-important">
       <thead>
         <tr>
           <th>Date</th>
