+++ query-connector/src/app/query/components/resultsView/tableComponents/ObservationTable.tsx
@@ -19,7 +19,7 @@ const ObservationTable: React.FC<ObservationTableProps> = ({
   observations,
 }) => {
   return (
-    <Table>
+    <Table className="margin-top-0-important">
       <thead>
         <tr>
           <th>Date</th>
