--- query-connector/src/app/query/page.module.css
@@ -1,4 +1,0 @@
-.resultsBannerContent {
-  display: flex;
-  align-items: center;
-}
